+++ ValheimArmory/VAConfig.cs
@@ -1,4 +1,5 @@
 ﻿using BepInEx.Configuration;
+using ValheimArmory.common;
 
 namespace ValheimArmory
 {
@@ -8,6 +9,9 @@ namespace ValheimArmory
         public static ConfigEntry<bool> EnableDebugMode;
         public static ConfigEntry<float> HybridWeaponBloodMagicSkillIncrease;
         public static ConfigEntry<bool> VanillaHammersHavePrimaryAttack;
+        public static ConfigEntry<float> StagbreakerPrimaryAttackStamina;
+        public static ConfigEntry<float> IronSledgePrimaryAttackStamina;
+        public static ConfigEntry<float> DemolisherPrimaryAttackStamina;
         public VAConfig(ConfigFile Config)
         {
             // ensure all the config values are created
@@ -27,6 +31,12 @@ namespace ValheimArmory
                 new ConfigurationManagerAttributes { IsAdvanced = true }));
             HybridWeaponBloodMagicSkillIncrease = BindServerConfig("Blood Magic Hybrid Weapons", "HybridWeaponBloodMagicSkillIncrease", 1f, "How much experiance should one usage of a blood magic hybrid weapon provide?", true, 0f, 4f);
             VanillaHammersHavePrimaryAttack = BindServerConfig("Vanilla Weapons", "VanillaHammersHavePrimaryAttack", true, "Enables a primary swing for vanilla sledges. Moves the slam to a secondary attack.");
+            StagbreakerPrimaryAttackStamina = BindServerConfig("Vanilla Weapons", "StagbreakerPrimaryAttackStamina", 6f, "Stamina cost of the basic attack when enabled for the stagbreaker.", true, 1, 30);
+            StagbreakerPrimaryAttackStamina.SettingChanged += WeaponModifier.OnConfigStagbreakerValueChanged;
+            IronSledgePrimaryAttackStamina = BindServerConfig("Vanilla Weapons", "IronSledgePrimaryAttackStamina", 10f, "Stamina cost of the basic attack when enabled for the iron sledge.", true, 1, 30);
+            IronSledgePrimaryAttackStamina.SettingChanged += WeaponModifier.OnConfigIronSledgeValueChanged;
+            DemolisherPrimaryAttackStamina = BindServerConfig("Vanilla Weapons", "DemolisherPrimaryAttackStamina", 14f, "Stamina cost of the basic attack when enabled for the demolisher.", true, 1, 30);
+            DemolisherPrimaryAttackStamina.SettingChanged += WeaponModifier.OnConfigDemolisherValueChanged;
         }
 
         /// <summary>