+++ ValheimArmory/ValheimArmory.cs
@@ -24,7 +24,7 @@ namespace ValheimArmory
     {
         public const string PluginGUID = "MidnightsFX.ValheimArmory";
         public const string PluginName = "ValheimArmory";
-        public const string PluginVersion = "1.16.0";
+        public const string PluginVersion = "1.16.2";
 
         internal static AssetBundle EmbeddedResourceBundle;
         CustomLocalization Localization;
@@ -47,7 +47,10 @@ namespace ValheimArmory
             // UnloadAssets();
 
             // Modify vanilla weapons
-            PrefabManager.OnVanillaPrefabsAvailable += WeaponModifier.ModifyVanillaHammersToWarhammers;
+            PrefabManager.OnVanillaPrefabsAvailable += WeaponModifier.SetupEffects;
+            if (VAConfig.VanillaHammersHavePrimaryAttack.Value) {
+                PrefabManager.OnVanillaPrefabsAvailable += WeaponModifier.ModifyVanillaHammersToWarhammers;
+            }
             VAConfig.VanillaHammersHavePrimaryAttack.SettingChanged += WeaponModifier.OnConfigChangeModifyHammers;
 
             Assembly assembly = Assembly.GetExecutingAssembly();